+++ bitcoin-de-sdk-value/src/main/kotlin/de/nykon/bitcoin/sdk/value/showOrderbookCompact/Ask.kt
@@ -1,6 +1,8 @@
 package de.nykon.bitcoin.sdk.value.showOrderbookCompact
 
+import java.math.BigDecimal
+
 data class Ask(
-    val amount_currency_to_trade: Double,
-    val price: Double
+        val amount_currency_to_trade: BigDecimal,
+        val price: BigDecimal
 )
